Split GJK module
