+++ distance3d/gjk/__init__.py
@@ -1,0 +1,6 @@
+"""Gilbert-Johnson-Keerthi (GJK) for distance calculation of convex shapes."""
+from .gjk_distance import gjk, gjk_with_simplex
+from .gjk_intersection import gjk_intersection
+
+
+__all__ = ["gjk", "gjk_with_simplex", "gjk_intersection"]
